Adding a Wrapper for the logger.
